Rename offspec_parameters_dialog to smooth_dialog for cleaner git diff
