+++ src/quicknxs/ui/ui_smooth_dialog.ui
